+++ tactigon_shapes/models.py
@@ -23,31 +23,26 @@ import json
 from os import getcwd
 from datetime import datetime
 from dataclasses import dataclass, field
-from typing import Optional, List
 
 from tactigon_gear import  __version__ as tactigon_gear_version
-
-if sys.platform != "darwin":
-    from tactigon_speech import __version__ as tactigon_speech_version
-else:
-    tactigon_speech_version = None
-
-from .modules.tskin.models import Hand, TSkinModel, TSkinConfig, VoiceConfig, ModelGesture, ModelTouch, OneFingerGesture
+from .modules.tskin.models import Hand, TSkinModel, Scorer, TSkinConfig, SocketConfig, ModelGesture, ModelTouch, OneFingerGesture, TSpeechObject
 
 BASE_PATH = getcwd()
 
 TACTIGON_GEAR = tactigon_gear_version
-TACTIGON_SPEECH = tactigon_speech_version
 
 @dataclass
 class AppConfig(object):
     DEBUG: bool
     SECRET_KEY: str
     SEND_FILE_MAX_AGE_DEFAULT: int = 1
-    MODELS: List[TSkinModel] = field(default_factory=list)
-    TSKIN: Optional[TSkinConfig] = None
-    TSKIN_VOICE: Optional[VoiceConfig] = None
-    file_path: Optional[str] = None
+    MODELS: list[TSkinModel] = field(default_factory=list)
+    SCORERS: list[Scorer] = field(default_factory=list)
+    TSKIN: TSkinConfig | None = None
+    TSKIN_SOCKET: SocketConfig | None = None
+    SELECTED_SCORER: str | None = None
+    TSKIN_SPEECH: TSpeechObject | None = None
+    file_path: str | None = None
 
     @classmethod
     def Default(cls, file_path):
@@ -65,6 +60,9 @@ class AppConfig(object):
             ModelTouch(OneFingerGesture.SINGLE_TAP, "Tap"),
             ModelTouch(OneFingerGesture.TAP_AND_HOLD, "Tap and hold"),
         ]
+        scorer = [
+            Scorer("shapes", "shapes.scorer", "shapes.json")
+        ]
         return cls(
             DEBUG=True,
             SECRET_KEY="change-me",
@@ -72,6 +70,7 @@ class AppConfig(object):
                 TSkinModel("MODEL_01_LEFT", Hand.LEFT, datetime.now(), gestures, touchs),
                 TSkinModel("MODEL_01_RIGHT", Hand.RIGHT, datetime.now(), gestures, touchs),
             ],
+            SCORERS=scorer,
             file_path=file_path
         )
 
@@ -86,9 +85,12 @@ class AppConfig(object):
             json["DEBUG"],
             json["SECRET_KEY"],
             json["SEND_FILE_MAX_AGE_DEFAULT"],
-            [TSkinModel.FromJSON(f) for f in json["MODELS"]],
+            [TSkinModel.FromJSON(f) for f in json.get("MODELS", [])],
+            [Scorer.FromJSON(s) for s in json.get("SCORERS", [])],
             TSkinConfig.FromJSON(json["TSKIN"]) if "TSKIN" in json and json["TSKIN"] is not None else None,
-            VoiceConfig.FromJSON(json["TSKIN_VOICE"]) if "TSKIN_VOICE" in json and json["TSKIN_VOICE"] is not None and sys.platform != "darwin" else None,
+            SocketConfig.FromJSON(json["TSKIN_SOCKET"]) if "TSKIN_SOCKET" in json and json["TSKIN_SOCKET"] is not None else None,
+            json.get("SELECTED_SCORER", None),
+            TSpeechObject.FromJSON(json["TSKIN_SPEECH"]) if "TSKIN_SPEECH" in json and json["TSKIN_SPEECH"] is not None else None,
             file_path
             )
     
@@ -98,8 +100,11 @@ class AppConfig(object):
             "SECRET_KEY": self.SECRET_KEY,
             "SEND_FILE_MAX_AGE_DEFAULT": self.SEND_FILE_MAX_AGE_DEFAULT,
             "MODELS": [m.toJSON() for m in self.MODELS],
+            "SCORERS": [s.toJSON() for s in self.SCORERS],
             "TSKIN": self.TSKIN.toJSON() if self.TSKIN else None,
-            "TSKIN_VOICE": self.TSKIN_VOICE.toJSON() if self.TSKIN_VOICE else None
+            "TSKIN_SOCKET": self.TSKIN_SOCKET.toJSON() if self.TSKIN_SOCKET else None,
+            "SELECTED_SCORER": self.SELECTED_SCORER,
+            "TSKIN_SPEECH": self.TSKIN_SPEECH.toJSON() if self.TSKIN_SPEECH else None,
         }
     
     def save(self):