+++ tactigon_shapes/modules/shapes/blueprint.py
@@ -21,28 +21,23 @@
 import json
 from uuid import UUID, uuid4
 from datetime import datetime
-from typing import List, Optional
 
-from anyio import Condition
-from flask import Blueprint, render_template, flash, redirect, url_for
+from flask import Blueprint, render_template, flash, redirect, url_for,request, send_file
 
+from tactigon_shapes.config import app_config, check_config, allowed_extensions
+from tactigon_shapes.models import ModelGesture
+from tactigon_shapes.modules.shapes.extension import ShapeConfig, Program
+from tactigon_shapes.modules.shapes.manager import get_shapes_app
 from tactigon_shapes.modules.ginos.models import GinosConfig
-from tactigon_shapes.modules.mqtt.models import MQTTConfig
-
-from ..ironboy.manager import get_ironboy_interface
-
-from .extension import ShapeConfig, Program
-from .manager import get_shapes_app
-
-from ..tskin.manager import get_tskin
-from ..zion.manager import get_zion_interface
-
-from ..ginos.manager import get_ginos_blocks
-from ..mqtt.models import MQTTSubscription
-
-from ...config import app_config, check_config
-from ...models import ModelGesture
-from ...utils.request_utils import get_from_request, check_empty_inputs
+from tactigon_shapes.modules.mqtt.models import MQTTConfig, MQTTSubscription
+from tactigon_shapes.modules.ironboy.manager import get_ironboy_interface
+from tactigon_shapes.modules.tskin.manager import get_tskin
+from tactigon_shapes.modules.zion.manager import get_zion_interface
+from tactigon_shapes.modules.ginos.manager import get_ginos_blocks
+from tactigon_shapes.modules.ros2.extension import Ros2Interface
+from tactigon_shapes.modules.ros2.models import Ros2Subscription, Ros2Publisher, Ros2ShapeConfig
+from tactigon_shapes.modules.ros2.manager import get_ros2_interface
+from tactigon_shapes.utils.request_utils import get_from_request
 
 
 bp = Blueprint("shapes", __name__, url_prefix="/shapes", template_folder="templates", static_folder="static")
@@ -50,14 +45,14 @@ bp = Blueprint("shapes", __name__, url_prefix="/shapes", template_folder="templa
 @bp.route("/")
 @bp.route("/<string:program_id>")
 @check_config
-def index(program_id: Optional[str] = None):
+def index(program_id: str | None = None):
     _shapes = get_shapes_app()
 
     if not _shapes:
         flash("Shapes app not found", category="danger")
         return redirect(url_for("main.index"))
 
-    current_config: Optional[ShapeConfig] = None
+    current_config: ShapeConfig | None = None
 
     if program_id:
         program = _shapes.find_shape_by_id(UUID(program_id))
@@ -71,7 +66,7 @@ def index(program_id: Optional[str] = None):
         if any(_shapes.config):
             current_config = _shapes.config[0]
 
-    gesture_list: List[ModelGesture] = []
+    gesture_list: list[ModelGesture] = []
 
     if app_config.TSKIN and app_config.TSKIN.gesture_config:
         for model in app_config.MODELS:
@@ -81,18 +76,20 @@ def index(program_id: Optional[str] = None):
 
     blocks_config = _shapes.get_blocks_congfig(gesture_list)
 
-    if app_config.TSKIN_VOICE and app_config.TSKIN_VOICE.voice_commands:
-        blocks_config["speechs"] = _shapes.get_speech_block_config(app_config.TSKIN_VOICE.voice_commands)
+    if app_config.TSKIN_SPEECH:
+        blocks_config["speechs"] = _shapes.get_speech_block_config(app_config.TSKIN_SPEECH)
 
     zion = get_zion_interface()
-
     if zion and zion.devices:
         blocks_config["zion"] = zion.get_shape_blocks()
 
     ironboy = get_ironboy_interface()
-
     if ironboy:
         blocks_config["ironboy"] = ironboy.get_shape_blocks()
+
+    ros2_interface = get_ros2_interface()
+    if ros2_interface:
+        blocks_config["ros2"] = ros2_interface.get_blocks()
 
     blocks_config["ginos"] = get_ginos_blocks()
     
@@ -201,8 +198,7 @@ def edit(program_id: str):
         return redirect(url_for("shapes.index"))
     
     state = _shapes.get_state(current_config.id)
-
-    gesture_list: List[ModelGesture] = []
+    gesture_list: list[ModelGesture] = []
 
     if app_config.TSKIN and app_config.TSKIN.gesture_config:
         for model in app_config.MODELS:
@@ -212,11 +208,11 @@ def edit(program_id: str):
 
     blocks_config = _shapes.get_blocks_congfig(gesture_list)
 
-    if app_config.TSKIN_VOICE and app_config.TSKIN_VOICE.voice_commands:
-        blocks_config["speechs"] = _shapes.get_speech_block_config(app_config.TSKIN_VOICE.voice_commands)
+    if app_config.TSKIN_SPEECH:
+        blocks_config["speechs"] = _shapes.get_speech_block_config(app_config.TSKIN_SPEECH)
 
     zion = get_zion_interface()
-
+    ros2_interface = get_ros2_interface()
     ironboy = get_ironboy_interface()
 
     if ironboy:
@@ -224,6 +220,9 @@ def edit(program_id: str):
 
     if zion and zion.devices:
         blocks_config["zion"] = zion.get_shape_blocks()
+
+    if ros2_interface:
+        blocks_config["ros2"] = ros2_interface.get_blocks()
 
     blocks_config["ginos"] = get_ginos_blocks()
 
@@ -363,15 +362,16 @@ def clone_config(program_id: str):
         return redirect(url_for("shapes.index", program_id=program_id))
     
     new_config = ShapeConfig(
-        uuid4(),
-        program_name,
-        datetime.now(),
-        datetime.now(),
-        program_description,
-        False,
-        original_config.app_file,
-        original_config.ginos_config,
-        original_config.mqtt_config
+        id=uuid4(),
+        name=program_name,
+        created_on=datetime.now(),
+        modified_on=datetime.now(),
+        description=program_description,
+        readonly=False,
+        app_file=original_config.app_file,
+        ginos_config=original_config.ginos_config,
+        mqtt_config=original_config.mqtt_config,
+        ros2_config=original_config.ros2_config,
     )
 
     program = _shapes.get_shape(UUID(program_id))
@@ -397,6 +397,8 @@ def save_program(program_id: str):
     code = get_from_request('generatedCode')
     state = get_from_request('state')
     subscriptions = get_from_request("subscriptions")
+    ros2_publishers = get_from_request("ros2_publishers")
+    ros2_subscriptions = get_from_request("ros2_subscriptions")
 
     # is_empty_input = check_empty_inputs(locals().items())
 
@@ -414,7 +416,15 @@ def save_program(program_id: str):
     config.modified_on = datetime.now()
     
     if subscriptions and config.mqtt_config:
-        config.mqtt_config.subscriptions = [MQTTSubscription.FromJSON(s) for s in json.loads(subscriptions)]   
+        config.mqtt_config.subscriptions = [MQTTSubscription.FromJSON(s) for s in json.loads(subscriptions)]
+
+    if ros2_publishers or ros2_subscriptions:
+        ros2_config = Ros2ShapeConfig(
+            config.name,
+            [Ros2Publisher.FromJSON(p) for p in json.loads(ros2_publishers)] if ros2_publishers else [],
+            [Ros2Subscription.FromJSON(s) for s in json.loads(ros2_subscriptions)] if ros2_subscriptions else [],
+        )
+        config.ros2_config = ros2_config
 
     is_success = _shapes.update(config, Program(code=code, state=json.loads(state)))
 
@@ -504,5 +514,57 @@ def delete(program_id: str):
 
     _shapes.remove(UUID(program_id))
 
-    flash(f"Shape deleted.", category="success")
+    flash("Shape deleted.", category="success")
     return redirect(url_for("shapes.index"))
+
+@bp.route("/import", methods=["POST"])
+@check_config
+def import_shape():
+    _shapes = get_shapes_app()
+
+    if not _shapes:
+        flash(f"Shapes app not found!", category="danger")
+        return redirect(url_for("main.index"))
+    
+    zip_file = request.files['file']
+
+    if not zip_file or not zip_file.filename:
+        flash("Shape file not found!", category="danger")
+        return redirect(url_for("shapes.index"))
+        
+    if not zip_file.filename.endswith(tuple(allowed_extensions)):
+        flash(f"Invalid file type. Allowed types: {', '.join(allowed_extensions)}", category="danger")
+        return redirect(url_for("shapes.index"))
+    
+    config, message =_shapes.import_shape(zip_file)
+
+    if not config:
+        flash(message, category="danger")
+        return redirect(url_for("shapes.index"))
+
+    flash(message, category="success")
+    return redirect(url_for("shapes.edit", program_id=config.id))
+
+@bp.route("/<string:program_id>/export")
+@check_config
+def export_shape(program_id: str):
+    _shapes = get_shapes_app()
+
+    if not _shapes:
+        flash("Shapes app not found!", category="danger")
+        return redirect(url_for("main.index"))
+
+    config = _shapes.find_shape_by_id(UUID(program_id))
+
+    if not config:
+        flash("Shape not found!", category="danger")
+        return redirect(url_for("shapes.index"))
+
+    exported_file = _shapes.export(config)
+
+    return send_file(
+        exported_file,
+        mimetype="application/zip",
+        as_attachment=True,
+        download_name=f"{config.name}_export.zip"
+    )