+++ tactigon_shapes/modules/tskin/models.py
@@ -23,142 +23,17 @@ import json
 from os import path
 from datetime import datetime
 from dataclasses import dataclass, field
-from typing import Optional, List, Tuple, Iterable
 
-from tactigon_gear import TSkinConfig, GestureConfig, Gesture, Hand, Angle, Touch, OneFingerGesture, TwoFingerGesture
+from tactigon_gear import TSkinSocket as TSkin, TSkinConfig, GestureConfig, SocketConfig
+from tactigon_gear.models.tskin import Gesture, Hand, Angle, Touch, OneFingerGesture, TwoFingerGesture
+from tactigon_gear.models.audio import TSpeechObject, TSpeech, HotWord
 
-if sys.platform != "darwin":
-    from tactigon_speech import TSkin_Speech as OldTSkin, VoiceConfig as OldVoiceConfig, Transcription, TSpeech, TSpeechObject, HotWord
-else:
-    from tactigon_gear import TSkin as OldTSkin
-    tactigon_speech_version = None
-
-if sys.platform != "darwin":
-    HotWords = Tuple[HotWord, Optional[Iterable["HotWords"]]]
-
-    @dataclass
-    class TranscriptionWithPath(Transcription):
-        @classmethod
-        def FromTranscription(cls, transcription: Transcription):
-            return TranscriptionWithPath(
-                transcription.text,
-                transcription.path,
-                transcription.time,
-                transcription.timeout
-            )
-
-        def filter_tree(self, tree: Optional[TSpeechObject]) -> Optional[TSpeechObject]:
-            def f(tree: Optional[TSpeechObject], tree_path: List[HotWord]) -> Optional[TSpeechObject]:
-                if not tree:
-                    return
-
-                if not tree_path:
-                    return tree
-
-                node, *rest = tree_path
-
-                branches = [branch.children for branch in tree.t_speech if node in branch.hotwords]
-
-                if not branches:
-                    return tree
-
-                filtered_tree, *_ = branches
-
-                return f(filtered_tree, rest)
-
-            return f(tree, self.path) if self.path else tree
-        
-        def __str__(self):
-            return F"TranscriptionWithPath {self.text}, Path: {', '.join([hw.word for hw in self.path] if self.path else [])}, Time: {self.time}, Timeout: {self.timeout}"
-        
-    @dataclass
-    class VoiceConfig(OldVoiceConfig):
-        voice_commands_notification: Optional[str] = path.join("config", "audio", "trigger.wav")
-        voice_commands: Optional[TSpeechObject] = None
-
-        stop_hotword = None # type: ignore
-
-        @classmethod
-        def FromJSON(cls, json: dict):
-            _t = super().FromJSON(json)
-            _t.stop_hotword = None # type: ignore
-            _t.voice_commands = TSpeechObject.FromJSON(json["voice_commands"]) if "voice_commands"in json and json["voice_commands"] else None
-            _t.voice_commands_notification = json["voice_commands_notification"] if "voice_commands_notification" in json else None
-            return _t
-        
-        def toJSON(self) -> dict:
-            d = super().toJSON()
-            d["voice_commands"] = self.voice_commands.toJSON() if self.voice_commands else None
-            d["voice_commands_notification"] = self.voice_commands_notification
-            return d
-
-    class TSkin(OldTSkin):
-        def __init__(self, config: TSkinConfig, voice_config: Optional[VoiceConfig], debug: bool = False):
-            if voice_config is None:
-                raise ValueError("Missing the configuration for the voice")
-            
-            OldTSkin.__init__(self, config, voice_config, debug=debug)
-
-        @property
-        def can_listen(self):
-            return True
-
-else:
-    @dataclass
-    class HotWord:
-        @classmethod
-        def FromJSON(cls, json: dict):
-            return cls()
-        
-        def toJSON(self) -> dict:
-            return {}
-        
-    HotWords = Tuple[HotWord, Optional[Iterable["HotWords"]]]
-
-    @dataclass
-    class TSpeech:
-        @classmethod
-        def FromJSON(cls, json: dict):
-            return cls()
-        
-        def toJSON(self) -> dict:
-            return {}
-        
-    @dataclass
-    class TSpeechObject:
-        t_speech: List[TSpeech] = field(default_factory=list)
-        @classmethod
-        def FromJSON(cls, json: dict):
-            return cls()
-        
-        def toJSON(self) -> dict:
-            return {}
-
-    @dataclass
-    class VoiceConfig:
-        voice_commands: Optional[TSpeechObject] = None
-        stop_hotword: Optional[HotWord] = None
-
-        @classmethod
-        def FromJSON(cls, json: dict):
-            return cls()
-        
-        def toJSON(self) -> dict:
-            return {}
-        
-    class TSkin(OldTSkin):
-        def __init__(self, config: TSkinConfig, voice: Optional[VoiceConfig], debug: bool = False):
-            OldTSkin.__init__(self, config, debug)
-
-        @property
-        def can_listen(self):
-            return False
 
 @dataclass
 class ModelGesture:
     gesture: str
     label: str
-    description: Optional[str] = None
+    description: str | None = None
 
     @classmethod
     def FromJSON(cls, json: dict):
@@ -178,7 +53,7 @@ class ModelGesture:
 class ModelTouch:
     gesture: OneFingerGesture
     label: str
-    description: Optional[str] = None
+    description: str | None = None
 
     @classmethod
     def FromJSON(cls, json: dict):
@@ -199,8 +74,8 @@ class TSkinModel:
     name: str
     hand: Hand
     date: datetime
-    gestures: List[ModelGesture]
-    touchs: List[ModelTouch]
+    gestures: list[ModelGesture]
+    touchs: list[ModelTouch]
 
     @classmethod
     def FromJSON(cls, json: dict):
@@ -219,4 +94,25 @@ class TSkinModel:
             "date": self.date.isoformat(),
             "gestures": [g.toJSON() for g in self.gestures],
             "touchs": [t.toJSON() for t in self.touchs]
+        }
+
+@dataclass
+class Scorer:
+    name: str
+    scorer_file: str
+    speech_file: str
+
+    @classmethod
+    def FromJSON(cls, json: dict):
+        return cls(
+            name=json["name"],
+            scorer_file=json["scorer_file"],
+            speech_file=json["speech_file"]
+        )
+    
+    def toJSON(self) -> dict:
+        return {
+            "name": self.name,
+            "scorer_file": self.scorer_file,
+            "speech_file": self.speech_file,
         }